always run the installer
previously we relied on the installer run with `--noop` to determine
whether a "real" run is required, but `--noop` doesn't properly report
changes right now (https://tickets.puppetlabs.com/browse/PUP-686) and
thus the installer was never run again after initial installation.

let's always run the installer instead and only report "changed" to
ansible when the installer exited with 2. this is slightly more
expensive than the old solution, as it will now run all hooks, even if
they're not strictly required.

this commit can be reverted when the installer gains support for proper
reporting of changes together with `--noop`.